chore(db): Insert blocks in chunks of desired per file blocks
I doubt this will have a noticeable performance impact, but
theoretically it is better and imo there's no impact on readability. If
anything a positive one, as instead of a random 1k chunk size, it's now
based on a related block constant.

Added a test for the error the existing comment mentions. I checked that
the error occurs when I increase the chunk size to 10k.